audit/requirements.txt
# Repository Comparison Audit System Requirements

# Core dependencies
PyYAML>=6.0
GitPython>=3.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0

# Report generation
jinja2>=3.0.0
markdown>=3.0.0
